renaming user service file
